move component to folder components/ServiceCard
